# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VIRALOS AI - Autonomous Viral Content Operating System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /voiceovers/generate:
    post:
      operationId: generateVoiceover
      tags: [voiceovers]
      summary: Generate a voiceover for a script
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoiceoverInput"
      responses:
        "200":
          description: Voiceover job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceoverJob"
  /voiceovers/{projectId}:
    get:
      operationId: getVoiceover
      tags: [voiceovers]
      summary: Get voiceover for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Voiceover
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceoverJob"

  # ── Captions ─────────────────────────────────────────────────────────────
...
  /videos/render:
    post:
      operationId: renderVideo
      tags: [videos]
      summary: Trigger automated video editing and render
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenderInput"
      responses:
        "200":
          description: Render job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RenderJob"
  /videos/{projectId}/status:
    get:
      operationId: getVideoStatus
      tags: [videos]
      summary: Get render status for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Render job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RenderJob"
  /videos/{projectId}:
    get:
      operationId: getVideo
      tags: [videos]
      summary: Get the final video for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"

  # ── Exports ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    VoiceoverInput:
      type: object
      required: [projectId, text]
      properties:
        projectId:
          type: integer
        text:
          type: string
        voiceStyle:
          type: string
          enum: [motivational_male, motivational_female, dramatic_male, dramatic_female, calm_male, calm_female]
        emotionalTone:
          type: string
    VoiceoverJob:
      type: object
      required: [id, projectId, status, voiceStyle, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        status:
          type: string
          enum: [queued, processing, done, failed]
        voiceStyle:
          type: string
        audioUrl:
          type: ["string", "null"]
        durationMs:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
...
    RenderInput:
      type: object
      required: [projectId]
      properties:
        projectId:
          type: integer
        transitionStyle:
          type: string
          enum: [cinematic, fast_cut, smooth, beat_sync]
        backgroundMusic:
          type: boolean
        autoZoom:
          type: boolean
    RenderJob:
      type: object
      required: [id, projectId, status, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        status:
          type: string
          enum: [queued, processing, done, failed]
        progress:
          type: integer
        stage:
          type: ["string", "null"]
        outputUrl:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    Video:
      type: object
      required: [id, projectId, url, platform, durationSec, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        url:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        platform:
          type: string
        durationSec:
          type: integer
        fileSizeBytes:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
...