# lib/api-zod/src/generated/types/video.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VIRALOS AI - Autonomous Viral Content Operating System
 * OpenAPI spec version: 0.1.0
 */

export interface Video {
  id: number;
  projectId: number;
  url: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  platform: string;
  durationSec: number;
  /** @nullable */
  fileSizeBytes?: number | null;
  createdAt: Date;
}